refactor(github-issues-runtime-tests): extract command workflow module (#1268)
